@@ -11,59 +11,55 @@
     "test:coverage": "npm run test:headless -- --code-coverage",
     "lint": "ng lint angular8-yandex-maps",
     "release": "standard-version",
-    "prepare": "husky install",
-    "pre-commit": "lint-staged"
-  },
-  "lint-staged": {
-    "*.{js,ts,json,md,html,css}": "prettier --write",
-    "*.{js,ts}": "eslint --fix"
+    "prepare": "husky install"
   },
   "dependencies": {
-    "@angular/animations": "^13.2.3",
-    "@angular/common": "^13.2.3",
-    "@angular/compiler": "^13.2.3",
-    "@angular/core": "^13.2.3",
-    "@angular/forms": "^13.3.4",
-    "@angular/platform-browser": "^13.2.3",
-    "@angular/platform-browser-dynamic": "^13.3.4",
-    "@angular/router": "^13.3.4",
+    "@angular/animations": "^14.0.4",
+    "@angular/common": "^14.0.4",
+    "@angular/compiler": "^14.0.4",
+    "@angular/core": "^14.0.4",
+    "@angular/forms": "^14.0.4",
+    "@angular/platform-browser": "^14.0.4",
+    "@angular/platform-browser-dynamic": "^14.0.4",
+    "@angular/router": "^14.0.4",
     "rxjs": "^7.5.5",
-    "tslib": "^2.3.1",
-    "zone.js": "^0.11.5"
+    "tslib": "^2.4.0",
+    "zone.js": "^0.11.6"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^13.3.3",
-    "@angular-eslint/builder": "^13.2.1",
-    "@angular-eslint/eslint-plugin": "^13.1.0",
-    "@angular-eslint/eslint-plugin-template": "^13.1.0",
-    "@angular-eslint/schematics": "^13.2.1",
-    "@angular-eslint/template-parser": "^13.2.1",
-    "@angular/cli": "^13.3.3",
-    "@angular/compiler-cli": "^13.3.4",
+    "@angular-devkit/build-angular": "^14.0.4",
+    "@angular-eslint/builder": "^14.0.0",
+    "@angular-eslint/eslint-plugin": "^14.0.0",
+    "@angular-eslint/eslint-plugin-template": "^14.0.0",
+    "@angular-eslint/schematics": "^14.0.0",
+    "@angular-eslint/template-parser": "^14.0.0",
+    "@angular/cli": "^14.0.4",
+    "@angular/compiler-cli": "^14.0.4",
     "@compodoc/compodoc": "1.1.18",
-    "@types/jasmine": "^3.10.6",
-    "@types/node": "^16.11.27",
-    "@typescript-eslint/eslint-plugin": "^5.20.0",
-    "@typescript-eslint/parser": "^5.20.0",
-    "eslint": "^8.13.0",
+    "@types/jasmine": "^4.0.3",
+    "@types/node": "^18.0.0",
+    "@typescript-eslint/eslint-plugin": "^5.30.3",
+    "@typescript-eslint/parser": "^5.30.3",
+    "eslint": "^8.19.0",
     "eslint-config-airbnb-base": "^15.0.0",
-    "eslint-config-airbnb-typescript": "^16.2.0",
+    "eslint-config-airbnb-typescript": "^17.0.0",
     "eslint-config-prettier": "^8.5.0",
+    "eslint-plugin-simple-import-sort": "^7.0.0",
     "eslint-plugin-import": "^2.26.0",
-    "husky": "^7.0.4",
-    "jasmine-core": "~3.10.1",
+    "husky": "^8.0.1",
+    "jasmine-core": "~4.2.0",
     "jasmine-spec-reporter": "~7.0.0",
-    "karma": "^6.3.19",
-    "karma-chrome-launcher": "~3.1.0",
-    "karma-coverage": "~2.0.3",
-    "karma-jasmine": "~4.0.1",
-    "karma-jasmine-html-reporter": "^1.7.0",
-    "lint-staged": "^12.4.0",
-    "ng-packagr": "^13.3.1",
-    "prettier": "^2.6.2",
+    "karma": "^6.4.0",
+    "karma-chrome-launcher": "~3.1.1",
+    "karma-coverage": "~2.2.0",
+    "karma-jasmine": "~5.1.0",
+    "karma-jasmine-html-reporter": "^2.0.0",
+    "lint-staged": "^13.0.3",
+    "ng-packagr": "^14.0.2",
+    "prettier": "^2.7.1",
     "protractor": "~7.0.0",
-    "standard-version": "^9.3.2",
-    "ts-node": "~10.4.0",
-    "typescript": "~4.4.2"
+    "standard-version": "^9.5.0",
+    "ts-node": "~10.8.2",
+    "typescript": "~4.7.4"
   }
 }
