@@ -4,16 +4,16 @@ module.exports = {
     {
       files: ['*.ts'],
       parserOptions: {
-        project: ['tsconfig.json', 'e2e/tsconfig.json'],
-        createDefaultProgram: true,
+        project: ['tsconfig.eslint.json'],
       },
+      plugins: ['simple-import-sort'],
       extends: [
         'plugin:@angular-eslint/recommended',
         'plugin:@angular-eslint/template/process-inline-templates',
         'eslint:recommended',
-        'plugin:@typescript-eslint/recommended',
         'airbnb-base',
         'airbnb-typescript/base',
+        'plugin:@typescript-eslint/recommended',
         'prettier',
       ],
       rules: {
@@ -21,17 +21,9 @@ module.exports = {
         'import/no-extraneous-dependencies': 'off',
         'import/prefer-default-export': 'off',
         'no-console': ['error', { allow: ['warn', 'error'] }],
-        'no-underscore-dangle': 'off',
+        'simple-import-sort/imports': 'error',
+        'simple-import-sort/exports': 'error',
 
-        '@typescript-eslint/naming-convention': [
-          'error',
-          {
-            selector: 'memberLike',
-            modifiers: ['private'],
-            format: ['camelCase'],
-            leadingUnderscore: 'require',
-          },
-        ],
         '@typescript-eslint/no-explicit-any': 'off',
         '@typescript-eslint/no-non-null-assertion': 'off',
         '@typescript-eslint/prefer-readonly': ['error'],
@@ -40,7 +32,6 @@ module.exports = {
     {
       files: ['*.html'],
       extends: ['plugin:@angular-eslint/template/recommended'],
-      rules: {},
     },
   ],
 };
