@@ -3,9 +3,7 @@ import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { BehaviorSubject } from 'rxjs';
 
 import { AngularYandexMapsModule } from '../../angular-yandex-maps.module';
-import { YaClustererComponent } from './ya-clusterer.component';
-import { YaMapComponent } from '../ya-map/ya-map.component';
-import { YaReadyEvent } from '../../typings/ya-ready-event';
+import { YaReadyEvent } from '../../models/ya-ready-event';
 import {
   createClustererConstructorSpy,
   createClustererSpy,
@@ -15,6 +13,8 @@ import {
   createPlacemarkConstructorSpy,
   createPlacemarkSpy,
 } from '../../testing/fake-ymaps-utils';
+import { YaMapComponent } from '../ya-map/ya-map.component';
+import { YaClustererComponent } from './ya-clusterer.component';
 
 @Component({
   template: `
