@@ -9,12 +9,13 @@ import {
   Output,
   SimpleChanges,
 } from '@angular/core';
-import { Observable, Subscription } from 'rxjs';
+import { Observable, Subject, takeUntil } from 'rxjs';
+import { filter, take } from 'rxjs/operators';
 
-import { EventManager } from '../../event-manager';
-import { YaEvent } from '../../typings/ya-event';
+import { EventManager } from '../../event-manager/event-manager';
+import { YaEvent } from '../../models/ya-event';
+import { YaReadyEvent } from '../../models/ya-ready-event';
 import { YaMapComponent } from '../ya-map/ya-map.component';
-import { YaReadyEvent } from '../../typings/ya-ready-event';
 
 /**
  * @internal
@@ -42,11 +43,11 @@ type Optional<T, K extends keyof T> = Pick<Partial<T>, K> & Omit<T, K>;
   selector: 'ya-multiroute',
 })
 export class YaMultirouteDirective implements OnInit, OnChanges, OnDestroy {
-  private readonly _sub = new Subscription();
+  private readonly destroy$ = new Subject<void>();
 
-  private readonly _eventManager = new EventManager(this._ngZone);
+  private readonly eventManager = new EventManager(this.ngZone);
 
-  private _multiroute?: ymaps.multiRouter.MultiRoute;
+  private multiroute?: ymaps.multiRouter.MultiRoute;
 
   /**
    * Reference points for the multiroute.
@@ -80,166 +81,166 @@ export class YaMultirouteDirective implements OnInit, OnChanges, OnDestroy {
    * Change to the active route.
    */
   @Output() activeroutechange: Observable<YaEvent<ymaps.multiRouter.MultiRoute>> =
-    this._eventManager.getLazyEmitter('activeroutechange');
+    this.eventManager.getLazyEmitter('activeroutechange');
 
   /**
    * Closing the balloon.
    */
   @Output() balloonclose: Observable<YaEvent<ymaps.multiRouter.MultiRoute>> =
-    this._eventManager.getLazyEmitter('balloonclose');
+    this.eventManager.getLazyEmitter('balloonclose');
 
   /**
    * Opening a balloon on a map.
    */
   @Output() balloonopen: Observable<YaEvent<ymaps.multiRouter.MultiRoute>> =
-    this._eventManager.getLazyEmitter('balloonopen');
+    this.eventManager.getLazyEmitter('balloonopen');
 
   /**
    * The event occurs at the time of setting the map center and its zoom level for optimal display of the multi-route.
    */
   @Output() boundsautoapply: Observable<YaEvent<ymaps.multiRouter.MultiRoute>> =
-    this._eventManager.getLazyEmitter('boundsautoapply');
+    this.eventManager.getLazyEmitter('boundsautoapply');
 
   /**
    * Changing coordinates of the geographical area covering the multi-route.
    */
   @Output() boundschange: Observable<YaEvent<ymaps.multiRouter.MultiRoute>> =
-    this._eventManager.getLazyEmitter('boundschange');
+    this.eventManager.getLazyEmitter('boundschange');
 
   /**
    * Single left-click on the object.
    */
   @Output() yaclick: Observable<YaEvent<ymaps.multiRouter.MultiRoute>> =
-    this._eventManager.getLazyEmitter('click');
+    this.eventManager.getLazyEmitter('click');
 
   /**
    * Calls the element's context menu.
    */
   @Output() yacontextmenu: Observable<YaEvent<ymaps.multiRouter.MultiRoute>> =
-    this._eventManager.getLazyEmitter('contextmenu');
+    this.eventManager.getLazyEmitter('contextmenu');
 
   /**
    * Double left-click on the object.
    */
   @Output() yadblclick: Observable<YaEvent<ymaps.multiRouter.MultiRoute>> =
-    this._eventManager.getLazyEmitter('dblclick');
+    this.eventManager.getLazyEmitter('dblclick');
 
   /**
    * Change to the geo object geometry.
    */
   @Output() geometrychange: Observable<YaEvent<ymaps.multiRouter.MultiRoute>> =
-    this._eventManager.getLazyEmitter('geometrychange');
+    this.eventManager.getLazyEmitter('geometrychange');
 
   /**
    * Map reference changed.
    */
   @Output() mapchange: Observable<YaEvent<ymaps.multiRouter.MultiRoute>> =
-    this._eventManager.getLazyEmitter('mapchange');
+    this.eventManager.getLazyEmitter('mapchange');
 
   /**
    * Pressing the mouse button over the object.
    */
   @Output() yamousedown: Observable<YaEvent<ymaps.multiRouter.MultiRoute>> =
-    this._eventManager.getLazyEmitter('mousedown');
+    this.eventManager.getLazyEmitter('mousedown');
 
   /**
    * Pointing the cursor at the object.
    */
   @Output() yamouseenter: Observable<YaEvent<ymaps.multiRouter.MultiRoute>> =
-    this._eventManager.getLazyEmitter('mouseenter');
+    this.eventManager.getLazyEmitter('mouseenter');
 
   /**
    * Moving the cursor off of the object.
    */
   @Output() yamouseleave: Observable<YaEvent<ymaps.multiRouter.MultiRoute>> =
-    this._eventManager.getLazyEmitter('mouseleave');
+    this.eventManager.getLazyEmitter('mouseleave');
 
   /**
    * Moving the cursor over the object.
    */
   @Output() yamousemove: Observable<YaEvent<ymaps.multiRouter.MultiRoute>> =
-    this._eventManager.getLazyEmitter('mousemove');
+    this.eventManager.getLazyEmitter('mousemove');
 
   /**
    * Letting go of the mouse button over an object.
    */
   @Output() yamouseup: Observable<YaEvent<ymaps.multiRouter.MultiRoute>> =
-    this._eventManager.getLazyEmitter('mouseup');
+    this.eventManager.getLazyEmitter('mouseup');
 
   /**
    * End of multitouch.
    */
   @Output() multitouchend: Observable<YaEvent<ymaps.multiRouter.MultiRoute>> =
-    this._eventManager.getLazyEmitter('multitouchend');
+    this.eventManager.getLazyEmitter('multitouchend');
 
   /**
    * Repeating event during multitouch.
    */
   @Output() multitouchmove: Observable<YaEvent<ymaps.multiRouter.MultiRoute>> =
-    this._eventManager.getLazyEmitter('multitouchmove');
+    this.eventManager.getLazyEmitter('multitouchmove');
 
   /**
    * Start of multitouch.
    */
   @Output() multitouchstart: Observable<YaEvent<ymaps.multiRouter.MultiRoute>> =
-    this._eventManager.getLazyEmitter('multitouchstart');
+    this.eventManager.getLazyEmitter('multitouchstart');
 
   /**
    * Change to the object options.
    */
   @Output() optionschange: Observable<YaEvent<ymaps.multiRouter.MultiRoute>> =
-    this._eventManager.getLazyEmitter('optionschange');
+    this.eventManager.getLazyEmitter('optionschange');
 
   /**
    * Change to the geo object overlay.
    */
   @Output() overlaychange: Observable<YaEvent<ymaps.multiRouter.MultiRoute>> =
-    this._eventManager.getLazyEmitter('overlaychange');
+    this.eventManager.getLazyEmitter('overlaychange');
 
   /**
    * The parent object reference changed.
    */
   @Output() parentchange: Observable<YaEvent<ymaps.multiRouter.MultiRoute>> =
-    this._eventManager.getLazyEmitter('parentchange');
+    this.eventManager.getLazyEmitter('parentchange');
 
   /**
    * Changing pixel coordinates of the area covering the multi-route.
    */
   @Output() pixelboundschange: Observable<YaEvent<ymaps.multiRouter.MultiRoute>> =
-    this._eventManager.getLazyEmitter('pixelboundschange');
+    this.eventManager.getLazyEmitter('pixelboundschange');
 
   /**
    * Change to the geo object data.
    */
   @Output() propertieschange: Observable<YaEvent<ymaps.multiRouter.MultiRoute>> =
-    this._eventManager.getLazyEmitter('propertieschange');
+    this.eventManager.getLazyEmitter('propertieschange');
 
   /**
    * Updating the multi-route.
    */
   @Output() update: Observable<YaEvent<ymaps.multiRouter.MultiRoute>> =
-    this._eventManager.getLazyEmitter('update');
+    this.eventManager.getLazyEmitter('update');
 
   /**
    * Mouse wheel scrolling.
    */
   @Output() yawheel: Observable<YaEvent<ymaps.multiRouter.MultiRoute>> =
-    this._eventManager.getLazyEmitter('wheel');
+    this.eventManager.getLazyEmitter('wheel');
 
-  constructor(private readonly _ngZone: NgZone, private readonly _yaMapComponent: YaMapComponent) {}
+  constructor(private readonly ngZone: NgZone, private readonly yaMapComponent: YaMapComponent) {}
 
   /**
    * Handles input changes and passes them in API.
    * @param changes
    */
   ngOnChanges(changes: SimpleChanges): void {
-    const multiroute = this._multiroute;
+    const { multiroute } = this;
 
     if (multiroute) {
       const { referencePoints, model, options } = changes;
 
       if (model) {
-        this._setModel(model.currentValue, multiroute);
+        this.setModel(model.currentValue, multiroute);
       }
 
       if (referencePoints) {
@@ -254,29 +255,28 @@ export class YaMultirouteDirective implements OnInit, OnChanges, OnDestroy {
 
   ngOnInit(): void {
     // It should be a noop during server-side rendering.
-    if (this._yaMapComponent.isBrowser) {
-      const sub = this._yaMapComponent.map$.subscribe((map) => {
-        if (map) {
-          const multiroute = this._createMultiroute();
-          this._multiroute = multiroute;
+    if (this.yaMapComponent.isBrowser) {
+      this.yaMapComponent.map$
+        .pipe(filter(Boolean), take(1), takeUntil(this.destroy$))
+        .subscribe((map) => {
+          const multiroute = this.createMultiroute();
+          this.multiroute = multiroute;
 
           map.geoObjects.add(multiroute);
-          this._eventManager.setTarget(multiroute);
-          this._ngZone.run(() => this.ready.emit({ ymaps, target: multiroute }));
-        }
-      });
-
-      this._sub.add(sub);
+          this.eventManager.setTarget(multiroute);
+          this.ngZone.run(() => this.ready.emit({ ymaps, target: multiroute }));
+        });
     }
   }
 
   ngOnDestroy(): void {
-    if (this._multiroute) {
-      this._yaMapComponent?.map$.value?.geoObjects.remove(this._multiroute);
-      this._eventManager.destroy();
+    if (this.multiroute) {
+      this.yaMapComponent?.map$.value?.geoObjects.remove(this.multiroute);
+      this.eventManager.destroy();
     }
 
-    this._sub.unsubscribe();
+    this.destroy$.next();
+    this.destroy$.complete();
   }
 
   /**
@@ -284,7 +284,7 @@ export class YaMultirouteDirective implements OnInit, OnChanges, OnDestroy {
    * @param model
    * @param multiroute
    */
-  private _setModel(
+  private setModel(
     model: ymaps.IMultiRouteModelJson,
     multiroute: ymaps.multiRouter.MultiRoute,
   ): void {
@@ -302,14 +302,14 @@ export class YaMultirouteDirective implements OnInit, OnChanges, OnDestroy {
   /**
    * Creates Multiroute.
    */
-  private _createMultiroute(): ymaps.multiRouter.MultiRoute {
-    return new ymaps.multiRouter.MultiRoute(this._combineModel(), this.options);
+  private createMultiroute(): ymaps.multiRouter.MultiRoute {
+    return new ymaps.multiRouter.MultiRoute(this.combineModel(), this.options);
   }
 
   /**
    * Combines the model and reference points into single object
    */
-  private _combineModel(): ymaps.IMultiRouteModelJson {
+  private combineModel(): ymaps.IMultiRouteModelJson {
     const model = (this.model || {}) as ymaps.IMultiRouteModelJson;
 
     return {
