@@ -9,12 +9,13 @@ import {
   Output,
   SimpleChanges,
 } from '@angular/core';
-import { Observable, Subscription } from 'rxjs';
+import { Observable, Subject, takeUntil } from 'rxjs';
+import { filter, take } from 'rxjs/operators';
 
-import { EventManager } from '../../event-manager';
-import { YaEvent } from '../../typings/ya-event';
+import { EventManager } from '../../event-manager/event-manager';
+import { YaEvent } from '../../models/ya-event';
+import { YaReadyEvent } from '../../models/ya-ready-event';
 import { YaMapComponent } from '../ya-map/ya-map.component';
-import { YaReadyEvent } from '../../typings/ya-ready-event';
 
 /**
  * The `ya-object-manager` component wraps `ymaps.ObjectManager` class from the Yandex.Maps API.
@@ -37,11 +38,11 @@ import { YaReadyEvent } from '../../typings/ya-ready-event';
   selector: 'ya-object-manager',
 })
 export class YaObjectManagerDirective implements OnInit, OnChanges, OnDestroy {
-  private readonly _sub = new Subscription();
+  private readonly destroy$ = new Subject<void>();
 
-  private readonly _eventManager = new EventManager(this._ngZone);
+  private readonly eventManager = new EventManager(this.ngZone);
 
-  private _objectManager?: ymaps.ObjectManager;
+  private objectManager?: ymaps.ObjectManager;
 
   /**
    * Options for the ObjectManager.
@@ -60,118 +61,118 @@ export class YaObjectManagerDirective implements OnInit, OnChanges, OnDestroy {
    * Single left-click on the object.
    */
   @Output() yaclick: Observable<YaEvent<ymaps.ObjectManager>> =
-    this._eventManager.getLazyEmitter('click');
+    this.eventManager.getLazyEmitter('click');
 
   /**
    * Calls the element's context menu.
    */
   @Output() yacontextmenu: Observable<YaEvent<ymaps.ObjectManager>> =
-    this._eventManager.getLazyEmitter('contextmenu');
+    this.eventManager.getLazyEmitter('contextmenu');
 
   /**
    * Double left-click on the object.
    */
   @Output() yadblclick: Observable<YaEvent<ymaps.ObjectManager>> =
-    this._eventManager.getLazyEmitter('dblclick');
+    this.eventManager.getLazyEmitter('dblclick');
 
   /**
    * Change to the geo object geometry.
    */
   @Output() geometrychange: Observable<YaEvent<ymaps.ObjectManager>> =
-    this._eventManager.getLazyEmitter('geometrychange');
+    this.eventManager.getLazyEmitter('geometrychange');
 
   /**
    * Map reference changed.
    */
   @Output() mapchange: Observable<YaEvent<ymaps.ObjectManager>> =
-    this._eventManager.getLazyEmitter('mapchange');
+    this.eventManager.getLazyEmitter('mapchange');
 
   /**
    * Pressing the mouse button over the object.
    */
   @Output() yamousedown: Observable<YaEvent<ymaps.ObjectManager>> =
-    this._eventManager.getLazyEmitter('mousedown');
+    this.eventManager.getLazyEmitter('mousedown');
 
   /**
    * Pointing the cursor at the object.
    */
   @Output() yamouseenter: Observable<YaEvent<ymaps.ObjectManager>> =
-    this._eventManager.getLazyEmitter('mouseenter');
+    this.eventManager.getLazyEmitter('mouseenter');
 
   /**
    * Moving the cursor off of the object.
    */
   @Output() yamouseleave: Observable<YaEvent<ymaps.ObjectManager>> =
-    this._eventManager.getLazyEmitter('mouseleave');
+    this.eventManager.getLazyEmitter('mouseleave');
 
   /**
    * Moving the cursor over the object.
    */
   @Output() yamousemove: Observable<YaEvent<ymaps.ObjectManager>> =
-    this._eventManager.getLazyEmitter('mousemove');
+    this.eventManager.getLazyEmitter('mousemove');
 
   /**
    * Letting go of the mouse button over an object.
    */
   @Output() yamouseup: Observable<YaEvent<ymaps.ObjectManager>> =
-    this._eventManager.getLazyEmitter('mouseup');
+    this.eventManager.getLazyEmitter('mouseup');
 
   /**
    * End of multitouch.
    */
   @Output() multitouchend: Observable<YaEvent<ymaps.ObjectManager>> =
-    this._eventManager.getLazyEmitter('multitouchend');
+    this.eventManager.getLazyEmitter('multitouchend');
 
   /**
    * Repeating event during multitouch.
    */
   @Output() multitouchmove: Observable<YaEvent<ymaps.ObjectManager>> =
-    this._eventManager.getLazyEmitter('multitouchmove');
+    this.eventManager.getLazyEmitter('multitouchmove');
 
   /**
    * Start of multitouch.
    */
   @Output() multitouchstart: Observable<YaEvent<ymaps.ObjectManager>> =
-    this._eventManager.getLazyEmitter('multitouchstart');
+    this.eventManager.getLazyEmitter('multitouchstart');
 
   /**
    * Change to the object options.
    */
   @Output() optionschange: Observable<YaEvent<ymaps.ObjectManager>> =
-    this._eventManager.getLazyEmitter('optionschange');
+    this.eventManager.getLazyEmitter('optionschange');
 
   /**
    * Change to the geo object overlay.
    */
   @Output() overlaychange: Observable<YaEvent<ymaps.ObjectManager>> =
-    this._eventManager.getLazyEmitter('overlaychange');
+    this.eventManager.getLazyEmitter('overlaychange');
 
   /**
    * The parent object reference changed.
    */
   @Output() parentchange: Observable<YaEvent<ymaps.ObjectManager>> =
-    this._eventManager.getLazyEmitter('parentchange');
+    this.eventManager.getLazyEmitter('parentchange');
 
   /**
    * Change to the geo object data.
    */
   @Output() propertieschange: Observable<YaEvent<ymaps.ObjectManager>> =
-    this._eventManager.getLazyEmitter('propertieschange');
+    this.eventManager.getLazyEmitter('propertieschange');
 
   /**
    * Mouse wheel scrolling.
    */
   @Output() yawheel: Observable<YaEvent<ymaps.ObjectManager>> =
-    this._eventManager.getLazyEmitter('wheel');
+    this.eventManager.getLazyEmitter('wheel');
 
-  constructor(private readonly _ngZone: NgZone, private readonly _yaMapComponent: YaMapComponent) {}
+  constructor(private readonly ngZone: NgZone, private readonly yaMapComponent: YaMapComponent) {}
 
   /**
    * Handles input changes and passes them in API.
    * @param changes
    */
   ngOnChanges(changes: SimpleChanges): void {
-    const objectManager = this._objectManager;
+    const { objectManager } = this;
 
     if (objectManager) {
       const { options } = changes;
@@ -184,35 +185,34 @@ export class YaObjectManagerDirective implements OnInit, OnChanges, OnDestroy {
 
   ngOnInit(): void {
     // It should be a noop during server-side rendering.
-    if (this._yaMapComponent.isBrowser) {
-      const sub = this._yaMapComponent.map$.subscribe((map) => {
-        if (map) {
-          const objectManager = this._createObjectManager();
-          this._objectManager = objectManager;
+    if (this.yaMapComponent.isBrowser) {
+      this.yaMapComponent.map$
+        .pipe(filter(Boolean), take(1), takeUntil(this.destroy$))
+        .subscribe((map) => {
+          const objectManager = this.createObjectManager();
+          this.objectManager = objectManager;
 
           map.geoObjects.add(objectManager);
-          this._eventManager.setTarget(objectManager);
-          this._ngZone.run(() => this.ready.emit({ ymaps, target: objectManager }));
-        }
-      });
-
-      this._sub.add(sub);
+          this.eventManager.setTarget(objectManager);
+          this.ngZone.run(() => this.ready.emit({ ymaps, target: objectManager }));
+        });
     }
   }
 
   ngOnDestroy(): void {
-    if (this._objectManager) {
-      this._yaMapComponent?.map$.value?.geoObjects.remove(this._objectManager);
-      this._eventManager.destroy();
+    if (this.objectManager) {
+      this.yaMapComponent?.map$.value?.geoObjects.remove(this.objectManager);
+      this.eventManager.destroy();
     }
 
-    this._sub.unsubscribe();
+    this.destroy$.next();
+    this.destroy$.complete();
   }
 
   /**
    * Creates ObjectManager.
    */
-  private _createObjectManager(): ymaps.ObjectManager {
+  private createObjectManager(): ymaps.ObjectManager {
     return new ymaps.ObjectManager(this.options);
   }
 }
