@@ -1,5 +1,6 @@
-import { YaApiLoaderService, YaConfig } from './ya-api-loader.service';
+import { YaConfig } from '../../models/ya-config';
 import { createReadySpy } from '../../testing/fake-ymaps-utils';
+import { YaApiLoaderService } from './ya-api-loader.service';
 
 class FakeHTMLScriptElement {
   onHandlers: Record<string, any> = {};
