@@ -2,11 +2,12 @@
 
 import 'zone.js';
 import 'zone.js/testing';
+
+import { getTestBed } from '@angular/core/testing';
 import {
   BrowserDynamicTestingModule,
   platformBrowserDynamicTesting,
 } from '@angular/platform-browser-dynamic/testing';
-import { getTestBed } from '@angular/core/testing';
 
 declare const require: {
   context(
@@ -23,7 +24,9 @@ declare const require: {
 getTestBed().initTestEnvironment(BrowserDynamicTestingModule, platformBrowserDynamicTesting(), {
   teardown: { destroyAfterEach: false },
 });
+
 // Then we find all the tests.
 const context = require.context('./', true, /\.spec\.ts$/);
+
 // And load the modules.
 context.keys().map(context);
